refactoring for tests ver 1.0
