+++ src/test/java/com/epam/seleniumhw/mailru/tests/EmailDraftCreationTest.java
@@ -1,7 +1,6 @@
 package com.epam.seleniumhw.mailru.tests;
 
 
-import com.epam.seleniumhw.mailru.pageobject.MainPage;
 import com.epam.seleniumhw.mailru.utils.TestDataProvider;
 import io.qameta.allure.Epic;
 import io.qameta.allure.Feature;
@@ -11,8 +10,8 @@ import org.testng.annotations.Test;
 
 import java.util.List;
 
-import static com.epam.seleniumhw.mailru.utils.MailPartitionNameList.DRAFT;
-import static com.epam.seleniumhw.mailru.utils.TestHelper.getEmailListDataTestHelper;
+import static com.epam.seleniumhw.mailru.utils.MailTypeEnaum.DRAFT;
+import static com.epam.seleniumhw.mailru.utils.TestHelper.getStringEmailListFromWebElementList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Epic("Selenium_HW")
@@ -26,12 +25,7 @@ public class EmailDraftCreationTest extends BaseTest {
     public void createDraftEmailTest(String toWhomAddressEmailField, String subjectEmailField, String messageEmailField) {
 
         mainPage.createNewDraftEmail(toWhomAddressEmailField, subjectEmailField, messageEmailField);
-
-        List<String> listOfDraftUsers = getEmailListDataTestHelper(mainPage, DRAFT);
-
-        assertThat(listOfDraftUsers).as("Wrong mapping data!")
-                .isNotEmpty()
-                .contains(toWhomAddressEmailField);
+        mainPage.validateEmailListWithCurrentUser(toWhomAddressEmailField, DRAFT);
 
 
     }
