+++ src/test/java/com/epam/seleniumhw/mailru/utils/MailTypeEnaum.java
@@ -1,6 +1,6 @@
 package com.epam.seleniumhw.mailru.utils;
 
-public enum MailPartitionNameList {
+public enum MailTypeEnaum {
     DRAFT,
     SENT
 }
